@@ -8,54 +8,56 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "lib:build": "ng build ngb-filterable-dropdown"
+    "lib:build": "ng build ngb-filterable-dropdown",
+    "build-prod": "ng build ngb-filterable-dropdown --prod && cp README.md dist/@bugsplat/ngb-filterable-dropdown",
+    "release": "npm run build-prod && cd dist/@bugsplat/ngb-filterable-dropdown && npm publish"
   },
   "private": false,
   "license": "MIT",
   "repository": {
-    "type" : "git",
-    "url" : "https://github.com/BugSplat-Git/ngb-filterable-dropdown"
+    "type": "git",
+    "url": "https://github.com/BugSplat-Git/ngb-filterable-dropdown"
   },
   "bugs": {
     "url": "https://github.com/BugSplat-Git/ngb-filterable-dropdown/issues"
   },
   "dependencies": {
-    "@angular/animations": "~9.1.7",
-    "@angular/common": "~9.1.7",
-    "@angular/compiler": "~9.1.7",
-    "@angular/core": "~9.1.7",
-    "@angular/forms": "~9.1.7",
-    "@angular/platform-browser": "~9.1.7",
-    "@angular/platform-browser-dynamic": "~9.1.7",
-    "@angular/router": "~9.1.7",
-    "@ng-bootstrap/ng-bootstrap": "^5.0.0",
-    "bootstrap": "^4.3.1",
+    "@angular/animations": "~10.1.3",
+    "@angular/common": "~10.1.3",
+    "@angular/compiler": "~10.1.3",
+    "@angular/core": "~10.1.3",
+    "@angular/forms": "~10.1.3",
+    "@angular/platform-browser": "~10.1.3",
+    "@angular/platform-browser-dynamic": "~10.1.3",
+    "@angular/router": "~10.1.3",
+    "@ng-bootstrap/ng-bootstrap": "^7.0.0",
+    "bootstrap": "^4.5.2",
     "popper.js": "^1.16.1",
-    "rxjs": "~6.5.5",
-    "tslib": "^1.10.0",
+    "rxjs": "~6.6.3",
+    "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.6",
-    "@angular-devkit/build-ng-packagr": "~0.901.6",
-    "@angular/cli": "~9.1.6",
-    "@angular/compiler-cli": "~9.1.7",
-    "@angular/language-service": "~9.1.7",
+    "@angular-devkit/build-angular": "~0.1001.3",
+    "@angular-devkit/build-ng-packagr": "~0.1001.3",
+    "@angular/cli": "~10.1.3",
+    "@angular/compiler-cli": "~10.1.3",
+    "@angular/language-service": "~10.1.3",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.5.3",
-    "ng-packagr": "^9.0.0",
-    "protractor": "~5.4.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "ng-packagr": "^10.1.0",
+    "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.8.3"
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.3"
   }
 }
