@@ -1,6 +1,6 @@
 {
   "name": "@bugsplat/ngb-filterable-dropdown",
-  "version": "1.0.1",
+  "version": "2.0.0",
   "private": false,
   "license": "MIT",
   "repository": {
@@ -10,10 +10,13 @@
   "bugs": {
     "url": "https://github.com/BugSplat-Git/ngb-filterable-dropdown/issues"
   },
+  "dependencies": {
+    "tslib": "^2.0.0"
+  },
   "peerDependencies": {
-    "@angular/common": "^9.0.0",
-    "@angular/core": "^9.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^6.0.0",
-    "bootstrap": "^4.4.1"
+    "@angular/common": "^10.0.0",
+    "@angular/core": "^10.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^7.0.0",
+    "bootstrap": "^4.5.2"
   }
 }
