@@ -1,19 +1,19 @@
-import { CommonModule } from '@angular/common';
-import { async, ComponentFixture, fakeAsync, TestBed } from '@angular/core/testing';
-import { ReactiveFormsModule } from '@angular/forms';
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { take } from 'rxjs/operators';
-import { NgbFilterableDropdownComponent } from './ngb-filterable-dropdown.component';
-import { MultiSelectPipe } from './multi-select-pipe/multi-select-pipe';
+import { CommonModule } from "@angular/common";
+import { ComponentFixture, fakeAsync, TestBed } from "@angular/core/testing";
+import { ReactiveFormsModule } from "@angular/forms";
+import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
+import { take } from "rxjs/operators";
+import { MultiSelectPipe } from "./multi-select-pipe/multi-select-pipe";
+import { NgbFilterableDropdownComponent } from "./ngb-filterable-dropdown.component";
 
-describe('FilterableDropdownComponent', () => {
+describe("NgbFilterableDropdownComponent", () => {
   let component: NgbFilterableDropdownComponent;
   let fixture: ComponentFixture<NgbFilterableDropdownComponent>;
 
   let filterItem: string;
   let items: Array<string>;
 
-  beforeEach(async(() => {
+  beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
         NgbModule,
@@ -26,9 +26,7 @@ describe('FilterableDropdownComponent', () => {
       ]
     })
       .compileComponents();
-  }));
 
-  beforeEach(() => {
     fixture = TestBed.createComponent(NgbFilterableDropdownComponent);
     component = fixture.componentInstance;
     filterItem = "foo";
@@ -38,68 +36,67 @@ describe('FilterableDropdownComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should create', () => {
+  it("should create", () => {
     expect(component).toBeTruthy();
   });
 
-  //TODO ZL Talk about whether this should be actual functionality
-  it('should clear filter text when item is selected', () => {
+  it("should clear filter text when item is selected", () => {
     component.searchForm.controls["searchInput"].setValue(filterItem);
     component.onItemSelect(filterItem);
     expect(component.searchForm.controls["searchInput"].value).toEqual("");
   });
 
-  it('should populate filtered with correct list of filtered items', () => {
+  it("should populate filtered with correct list of filtered items", () => {
     component.searchInput.setValue("oo");
     expect(component.filtered.has(filterItem)).toEqual(true);
     expect(component.filtered.has("baz")).toEqual(false);
   });
 
-  it('should emit the value of the selected item', async () => {
+  it("should emit the value of the selected item", async () => {
     const resultPromise = component.onItemsSelected.pipe(take(1)).toPromise();
     component.onItemSelect(filterItem);
     const result = await resultPromise;
     expect(result).toEqual([filterItem])
   });
 
-  it('should display list of items', () => {
+  it("should display list of items", () => {
     fixture.detectChanges();
     items.forEach(item => expect(fixture.nativeElement.textContent).toContain(item));
   });
 
-  it('should be disabled if disabled input is true', () => {
+  it("should be disabled if disabled input is true", () => {
     component.disabled = true;
     fixture.detectChanges();
     expect(fixture.nativeElement.querySelector("button[disabled]")).toBeTruthy();
   });
 
-  it('should display All as currentItem if no items are selected', () => {
+  it("should display placeholder as currentItem if no items are selected", () => {
     component.items = ["it's a living"];
     fixture.detectChanges();
-    expect(fixture.nativeElement.querySelector('#itemDropdown').innerText).toEqual("All");
+    expect(fixture.nativeElement.querySelector("#toggle").innerText).toEqual(component.placeholder);
   });
 
-  it('should display currentItem if one item is selected', () => {
+  it("should display currentItem if one item is selected", () => {
     const currentItem = "MyDatabaseName";
     component.items = [currentItem];
     component.onItemSelect(currentItem);
     fixture.detectChanges();
-    expect(fixture.nativeElement.querySelector('#itemDropdown').innerText).toEqual(currentItem);
+    expect(fixture.nativeElement.querySelector("#toggle").innerText).toEqual(currentItem);
   });
 
-  it('should display Multiple as currentItem if more than one item is selected', fakeAsync(() => {
+  it("should display Multiple as currentItem if more than one item is selected", fakeAsync(() => {
     component.onItemSelect(filterItem);
     component.onItemSelect("bar");
     fixture.detectChanges();
-    expect(fixture.nativeElement.querySelector('#itemDropdown').innerText).toEqual("Multiple");
+    expect(fixture.nativeElement.querySelector("#toggle").innerText).toEqual("Multiple");
   }));
 
-  it('should emit an onOpen event when onOpenChange is called with true', () => {
+  it("should emit an onOpen event when onOpenChange is called with true", () => {
     component.onOpen.subscribe(() => expect(true).toBeTruthy());
     component.onOpenChange(true);
   });
 
-  it('should emit all items if "Select All" is selected', async () => {
+  it("should emit all items if 'Select All' is selected", async () => {
     const expectedResult = items.map(item => item);
     const resultPromise = component.onItemsSelected.pipe(take(1)).toPromise();
     component.disabled = false;
@@ -108,55 +105,55 @@ describe('FilterableDropdownComponent', () => {
     expect(result).toEqual(expectedResult)
   });
 
-  it('should emit an empty array if "Select None" is selected', () => {
+  it("should emit an empty array if 'Select None' is selected", () => {
     component.disabled = false;
     component.items = [];
     component.onItemsSelected.subscribe(items => expect(items).toEqual([]));
     component.onSelectNone();
   });
 
-  describe('isSelected', () => {
+  describe("isSelected", () => {
     beforeEach(() => {
       component.onItemSelect(filterItem);
     });
 
-    it('should should return true if item is selected', () => {
+    it("should should return true if item is selected", () => {
       expect(component.isSelected(filterItem)).toEqual(true)
     });
 
-    it('should set selected item when onItemSelect is called', () => {
+    it("should set selected item when onItemSelect is called", () => {
       expect(component.isSelected("baz")).toEqual(false);
     });
   });
 
-  describe('isFiltered', () => {
+  describe("isFiltered", () => {
     beforeEach(() => {
       component.searchInput.setValue(filterItem);
     });
 
-    it('should should return true if item is selected', () => {
+    it("should should return true if item is selected", () => {
       expect(component.isFiltered(filterItem)).toEqual(true)
     });
 
-    it('should set selected item when onItemSelect is called', () => {
+    it("should set selected item when onItemSelect is called", () => {
       expect(component.isFiltered("baz")).toEqual(false);
     });
   });
 
-  describe('noItemsToDisplay', () => {
-    it('should be true if filteredItems length is 0', () => {
+  describe("noItemsToDisplay", () => {
+    it("should be true if filteredItems length is 0", () => {
       component.disabled = false;
-      component.items = ['🍔'];
+      component.items = ["🍔"];
 
-      component.searchInput.setValue('alsdkjfals');
+      component.searchInput.setValue("alsdkjfals");
       fixture.detectChanges();
 
       expect(component.noItemsToDisplay).toEqual(true);
-      expect(fixture.nativeElement.querySelector('.no-items')).toBeTruthy();
+      expect(fixture.nativeElement.querySelector(".no-items")).toBeTruthy();
     });
 
-    it('should be false if filteredItems length is not 0', () => {
-      const item = '🍕';
+    it("should be false if filteredItems length is not 0", () => {
+      const item = "🍕";
       component.disabled = false;
       component.items = [item];
 
@@ -164,30 +161,50 @@ describe('FilterableDropdownComponent', () => {
       fixture.detectChanges();
 
       expect(component.noItemsToDisplay).toEqual(false);
-      expect(fixture.nativeElement.querySelector('.no-items')).toBeFalsy();
+      expect(fixture.nativeElement.querySelector(".no-items")).toBeFalsy();
     });
   });
 
-  describe('singleSelect', () => {
+  describe("singleSelect", () => {
     beforeEach(() => {
       component.allowMultiSelect = false;
     });
 
-    it('should set selected item', () => {
+    it("should set selected item", () => {
       component.selected = new Set([filterItem]);
       expect(component.selected.has(filterItem)).toEqual(true)
     });
 
-    it('should remove other items from set when onItemSelect is callsed', () => {
+    it("should remove other items from set when onItemSelect is callsed", () => {
       component.selected = new Set(["baz"]);
       component.onItemSelect(filterItem);
       expect(component.selected.has(filterItem)).toEqual(true);
       expect(component.selected.has("baz")).toEqual(false);
     })
 
-    it('should set selected item when onItemSelect is called', () => {
+    it("should set selected item when onItemSelect is called", () => {
       component.onItemSelect(filterItem);
       expect(component.selected.has(filterItem)).toEqual(true);
+    });
+  });
+
+  describe("autoClose", () => {
+    beforeEach(() => component.dropdown = jasmine.createSpyObj("NgbDropdown", ["close"]));
+
+    it("should close dialog when enter key is pressed if autoClose is true", () => {
+      component.autoClose = true;
+      
+      component.onEnterKeyPressed();
+
+      expect(component.dropdown.close).toHaveBeenCalled();
+    });
+
+    it("should not close dialog when enter key is pressed if autoClose is false", () => {
+      component.autoClose = false;
+      
+      component.onEnterKeyPressed();
+
+      expect(component.dropdown.close).not.toHaveBeenCalled();
     });
   });
 });
