@@ -1,40 +1,40 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { async, TestBed } from '@angular/core/testing';
-import { AppComponent } from './app.component';
+import { Component, EventEmitter, Input, Output } from "@angular/core";
+import { TestBed } from "@angular/core/testing";
+import { AppComponent } from "./app.component";
 
-describe('AppComponent', () => {
-  beforeEach(async(() => {
+describe("AppComponent", () => {
+  beforeEach(() => {
     TestBed.configureTestingModule({
       declarations: [
         AppComponent,
         MockFilterableDropdownComponent
       ],
     }).compileComponents();
-  }));
+  });
 
-  it('should create the app', () => {
+  it("should create the app", () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'filterable-dropdown-example'`, () => {
+  it("should have as title 'ngb-filterable-dropdown-example'", () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual('filterable-dropdown-example');
+    expect(app.title).toEqual("ngb-filterable-dropdown-example");
   });
 
-  it('should render title', () => {
+  it("should render title", () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('h2').textContent).toContain('filterable-dropdown-example');
+    expect(compiled.querySelector("h2").textContent).toContain("ngb-filterable-dropdown-example");
   });
 });
 
 @Component({
-  selector: 'bugsplat-filterable-dropdown',
-  template: ''
+  selector: "ngb-filterable-dropdown",
+  template: ""
 })
 class MockFilterableDropdownComponent {
   @Input() allowMultiSelect: any;
