@@ -1,19 +1,19 @@
-import { Component } from '@angular/core';
+import { Component } from "@angular/core";
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  selector: "app-root",
+  templateUrl: "./app.component.html",
+  styleUrls: ["./app.component.scss"]
 })
 export class AppComponent {
-  title = 'filterable-dropdown-example';
+  title = "ngb-filterable-dropdown-example";
   
-  readonly bugs: Array<string> = ['Beetle', 'Ant', 'Moth', 'Fire Ant', 'Dung Beetle', 'Grass Ant'] 
+  readonly bugs: Array<string> = ["Beetle", "Ant", "Moth", "Fire Ant", "Dung Beetle", "Grass Ant"] 
 
   allowMultiSelect: boolean = false;
-  autoClose: boolean | 'inside' | 'outside' = false;
+  autoClose: boolean | "inside" | "outside" = false;
   disabled: boolean = false;
-  selected: Array<string> = ['Moth'];
+  selected: Array<string> = ["Moth"];
 
   allowMultiSelectClick(event: CheckboxClickEvent): void {
     this.allowMultiSelect = event.target.checked;
@@ -21,7 +21,7 @@ export class AppComponent {
   }
 
   autoCloseClick(event: CheckboxClickEvent): void {
-    this.autoClose = event.target.checked ? 'inside' : false;
+    this.autoClose = event.target.checked ? "inside" : false;
   }
 
   disabledClick(event: CheckboxClickEvent): void {
@@ -29,7 +29,7 @@ export class AppComponent {
   }
 
   onDropdownOpen(): void {
-    console.log('Dropdown opened!');
+    console.log("Dropdown opened!");
   }
 
   onItemsSelected(items: string): void {
