@@ -1,5 +1,6 @@
 import { ChangeDetectorRef, Component, ElementRef, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
 import { AbstractControl, FormControl, FormGroup } from '@angular/forms';
+import { NgbDropdown } from '@ng-bootstrap/ng-bootstrap';
 import { SelectionType } from './selection-type';
 
 @Component({
@@ -26,13 +27,13 @@ export class NgbFilterableDropdownComponent implements OnInit {
   } 
   @Input() disabled: boolean = false;
   @Input() allowMultiSelect: boolean = true;
-  @Input() placeholder: string;
+  @Input() placeholder: string = 'No Items Selected';
 
   @Output() onItemsSelected: EventEmitter<Array<string> | string> = new EventEmitter<Array<string> | string>();
   @Output() onOpen: EventEmitter<void> = new EventEmitter<void>();
 
   @ViewChild('search', { static: true }) search: ElementRef;
-  @ViewChild('dropdown', { static: true }) dropdown: ElementRef;
+  @ViewChild('dropdown', { static: true }) dropdown: NgbDropdown;
 
   public get allowToggleSelectAll(): boolean {
     return (this.allowMultiSelect && this.searchInputValue.length === 0);
@@ -49,7 +50,6 @@ export class NgbFilterableDropdownComponent implements OnInit {
   });
 
   constructor(private changeDetector: ChangeDetectorRef) { }
-
 
   ngOnInit(): void {
     this.filtered = new Set(this.items);
@@ -70,13 +70,13 @@ export class NgbFilterableDropdownComponent implements OnInit {
     let arr: Array<any> = Array.from(this.selected);
     if (this.allowMultiSelect) {
       return arr;
-    } else {
-      if (arr) {
-        return arr[0]
-      } else {
-        return "";
-      }
     }
+
+    if (arr) {
+      return arr[0]
+    } 
+    
+    return "";
   }
 
   onSelectAll(): void {
@@ -93,7 +93,7 @@ export class NgbFilterableDropdownComponent implements OnInit {
     this.onItemsSelected.emit(this.selectedItems);
   }
 
-  onItemSelect(item: string) {
+  onItemSelect(item: string): void {
     if (this.allowMultiSelect) {
       if (this.selected.has(item)){
         this.selected.delete(item);
@@ -124,8 +124,13 @@ export class NgbFilterableDropdownComponent implements OnInit {
       });
     } else {
       if (this.filtered && this.filtered.size) {
-        this.selected = new Set([this.filtered[0]]);
+        this.selected = new Set([this.filtered.entries().next().value[0]]);
       } 
+    }
+
+    // TODO BG test
+    if (this.autoClose) {
+      this.dropdown.close();
     }
 
     this.onItemsSelected.emit(this.selectedItems);
